Logging improvements, refactor Run (#313)
