+++ fast_llm/engine/config_utils/logging.py
@@ -1,6 +1,10 @@
+import logging
 import logging.config
 import math
 import pathlib
+import typing
+
+logger = logging.getLogger(__name__)
 
 
 def configure_logging(
@@ -47,3 +51,35 @@ def configure_logging(
         }
         logging_config["root"]["handlers"].append("file")
     logging.config.dictConfig(logging_config)
+
+
+def log(*message, log_fn: typing.Union[BaseException, typing.Callable] = logger.info, join: str = ", "):
+    message = join.join([str(m() if callable(m) else m) for m in message])
+    if isinstance(log_fn, BaseException):
+        raise log_fn(message)
+    else:
+        return log_fn(message)
+
+
+class TensorLogs:
+    # A global buffer for holding logged tensor stats.
+    _tensor_log_stats: list | None = None
+    max_logged_elements = 8
+    verbose: bool = True
+
+    @classmethod
+    def reset(cls, enabled=True):
+        cls._tensor_log_stats = [] if enabled else None
+
+    @classmethod
+    def enabled(cls):
+        return cls._tensor_log_stats is not None
+
+    @classmethod
+    def append(cls, stats):
+        if cls._tensor_log_stats is not None:
+            cls._tensor_log_stats.append(stats)
+
+    @classmethod
+    def get(cls):
+        return cls._tensor_log_stats
