+++ fast_llm/engine/schedule/runner.py
@@ -9,12 +9,12 @@ import torch.cuda
 import yaml
 
 from fast_llm.core.distributed import all_reduce, recv, safe_barrier, send
+from fast_llm.engine.config_utils.run import get_run, log_pipeline_parallel_main_rank
 from fast_llm.engine.distributed.config import DistributedConfig
 from fast_llm.engine.distributed.distributed import Distributed
 from fast_llm.engine.multi_stage.multi_stage import MultiStageModel
 from fast_llm.engine.multi_stage.stage import Stage
 from fast_llm.engine.optimizer.optimizer import Optimizer
-from fast_llm.engine.run.run import log_pipeline_parallel_main_rank, open_artifact
 from fast_llm.engine.schedule.config import EventType, ScheduleConfig, StepType, StreamType
 from fast_llm.engine.schedule.schedule import Schedule, Step
 from fast_llm.logging import log_memory_usage
@@ -509,7 +509,7 @@ class ScheduleRunner:
         }
         yaml.safe_dump(
             out,
-            open_artifact(
+            get_run().open_artifact(
                 f"schedule_profile_rank_{self._distributed_config.rank}_{context.phase.value}_step_{context.iteration}"
             ),
         )
