+++ fast_llm/layers/common/normalization.py
@@ -1,7 +1,7 @@
 import torch
 
+from fast_llm.engine.config_utils.run import log_main_rank
 from fast_llm.engine.config_utils.tensor_space import TensorDim
-from fast_llm.engine.run.run import log_main_rank
 from fast_llm.functional.config import TritonConfig
 from fast_llm.functional.triton.normalization import rms_norm, triton_normalization_autograd
 from fast_llm.layers.common.config import NormalizationImplementation
