+++ fast_llm/tools/cli.py
@@ -3,7 +3,9 @@ import logging
 import sys
 import traceback
 
+from fast_llm.config import ValidationError
 from fast_llm.engine.config_utils.logging import configure_logging
+from fast_llm.engine.config_utils.run import log_main_rank
 
 logger = logging.getLogger(__name__)
 
@@ -26,6 +28,9 @@ def fast_llm(args=None):
             convert(unparsed)
         else:
             raise RuntimeError("Unknown subcommand")
+    except ValidationError:
+        log_main_rank(traceback.format_exc(), log_fn=logger.error)
+        sys.exit(1)
     except Exception:  # noqa
         logger.critical(traceback.format_exc())
         sys.exit(1)
